requirements.txt
celery==5.3.6
Django==5.0
PyMySQL==1.1.0
pytest==7.4.3
pytest-django==4.7.0
python-dotenv==0.20.0
redis==5.0.1